[gidl] Add tests for various recursive types
This CL replaces existing Go recursive types tests with a series of
tests covering a variety of recursive types. For unions and tables, it
is not possible today to have direct recursion so the tests cover a
"minimal" recursive type involving the target union/table type.

Bug: 35218
Change-Id: I2b65805332e1d53dd96410e90d839bd17401933f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606901
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>