[netcfg] Impl fuchsia.net.virtualization/Control
Allow clients to create bridged networks by keeping track of interfaces
that can provide upstream connectivity and bridging them with interfaces
that are added to the virtual network by the client.

Due to the limitations of the current bridging API, we add or remove
interfaces from the bridge by destroying the bridge and rebuilding it.

Fixed: 85078
Test: fx test netcfg-tests
Change-Id: I33cb399ff29ece2eba1506fd432fd3af8c521504
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595149
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tony Gong <gongt@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>