[ffx] ignore mdns packets with dubious addresses
When QEMU/AEMU are run using SLIRP they will write multicast packets out
to the host network approximately "as-is", however, those responses do
not necessarily indicate a viable return path. To date the code had
assumed that the host networking environment is sound with respect to
mdns propagation, but as it appears to not be so packets are now
validated against their source address.  If the packet does not contain
an answer that matches the source address it is now ignored, effectively
filtering out the NAT'd SLIRP targets. This may filter out desired
packets in some as yet not understood scenarios, in which case this
heuristic would need to be revised.

Fixed: 88587
Change-Id: I649d94e0a6b5e4c76a83128321f0ccda652f6403
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604646
Commit-Queue: James Tucker <raggi@google.com>
Fuchsia-Auto-Submit: James Tucker <raggi@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>