[brcmfmac] Replace scan_request with boolean scan_in_progress field
The pointer copied into the scan_request field of struct
brcmf_cfg80211_info is not owned by brcmfmac, so a proper copy of the
data it contains should be performed if we wish to retain it. Since
this field is only used as a boolean, instead of fixing the ownership
problem, the field is replaced by an actual boolean instead.

Fixed: 88807
Test: Pure refactor
Change-Id: I0c5ba63d780fc23546c3c06c40aa6b6a0a316d1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604987
Commit-Queue: Charles Celerier <chcl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>