[lib][concurrent] Start defining "libconcurrent"
Start defining "libconcurrent", a small source_set library intended to
contain a number of utilities which assist in safely and correctly
implementing synchronization patterns involving read operations
executing concurrently with write operations at the same memory
locations.

In this CL, we add the first building blocks in the form of
"WellDefinedCopyTo" and "WellDefinedCopyFrom".  These functions are
basically special forms of memcpy, but which use atomic stores (for
CopyTo) or loads (for CopyFrom) in order to avoid introducing data
races (which C++ considers to be Undefined Behavior) when reading from
and writing the same memory location(s) concurrent.

Change-Id: I1915dcb42da534ee107057a6204f3063e1b8ef4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602327
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>