[kernel][x86] Fix re-entrancy bug with swapgs and NMIs
This change fixes a bug where the kernel might use an invalid %gs.base
if an NMI interrupts another interrupt handler before that handler is
able to swapgs to the kernel's %gs.base.

Update interrupt_common to check whether the interrupt was an NMI and
if so save the current %gs.base on the stack and recover the kernel's
%gs.base (percpu pointer) using dead reckoning from the %rsp.

Add in-kernel unit test that would trigger the bug prior to this fix.

Bug: 31612
Bug: 86717
Bug: 87852
Bug: 88368
Test: k ut x64_platform_tests
Change-Id: I8d752e5fa880d34118afc0d7b6479019e1fdbeee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604194
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>