[netemul] Log guest kernel UART
When setting up a guest sandbox, fetch the guest's UART socket and log
it to the test's output. This will help diagnose issues in system
startup or guest kernel errors.

While this CL logs any serial that happens to be produced by guests, it
doesn't actually configure guests to produce serial output: this will be
done in an upcoming CL.

Change-Id: I0b0ddcbfddf7297bf29937267f97a5bb730a7fab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597982
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>