[rust-3p-crates] Bump rustls to 0.18.1
Updating rustls is a little challenging, since the latest version of
hyper-rustls and tokio-rustls depend on the latest versions of hyper and
tokio, which are quite large.

Fortunately we can reduce our load, because all the rustls dependencies
don't actually depend on any rustls breaking API changes. So this patch
takes rustls 0.18.1, updates it to incorporates the fix for
https://fxrev.dev/479224, and resets the version back to 0.17.0.

We also get to remove base64 0.11.0, but need to import base64 0.12.3.

Bug: 87798
Change-Id: Ia290b703159ba66fac560f1a4ba23448f12b6e36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601461
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Kostya Kortchinsky <kostyak@google.com>