[fx][pm] Warn when local and pm server incremental settings disagree
Adds a warning when the package server and shell from which an fx
command is run disagree on the incremental feature. This can cause
issues as the mechanism for how packages are published differ based on
the incremental feature.

For example, if 'fx serve' is run without the incremental option, and
'fx test' is run with the incremental option, the test will be rebuilt,
but the package is not published to the package server. This results in
the old version of the test being run with no warning to the developer.

Bug: 87471
Testing: Manually check the following conditions:
 1. Set incremental option in both serve-updates and fx test shells
 2. Set incremental option in only serve-updates
 3. Set incremental option in only fx test shell
 4. Don't set incremental option in either.
 Run 'fx test' and verify that only condition (3) results in a printed
 warning.

Change-Id: Ia186e7e683a12b9eb9695e5ac72b73968162b7c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607863
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>