[licenses] Skip directories at a lower level.
We shouldn't skip prebuilt/ and vendor/ directories entirely.
This CL aims to re-add as much of those sub-directories as possible.

In order to add more, we will need to be able to skip dirs with comments
in config files, so as to not leak details in fuchsia.git.
I'm working on that now.

Change-Id: Ic86db6743b4cca85fe7c90eb25640575cea80434
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608211
Reviewed-by: Sean Cuff <seancuff@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>