[f2fs] Add compatibility tests for extended inodes
If the filesystem is configured with extended inode features,
the allocated amount of on-disk bytes for inline dentries may decrease
and the way how a directory structures its inline dentries may differ.
According to that, this patch removes |struct InlineDentry| which is a
previously defined on-disk structure for inline dentries but rather
constructs instant pointers in data block pointers region when needed.

It also brings some minor bugfixes with extended inodes and adds
supports for configuring compatibility tests with Mkfs options.

Test: fx test f2fs-compatibility-test

Change-Id: I6730c2db6a3b664d7305620a59d93440acbbe73d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605881
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Chris Suter <csuter@google.com>