[ffx] Make sure ascendd binds the socket
This is a bit speculative from the failures we are seeing in test
flakes. My reading of the code is that there is technically a race
condition between telling ascendd to start up and bind a UDS and the
daemon trying to talk to that UDS. If the bind does not get scheduled in
time we could get the daemon to timeout trying to connect. It would look
like the failures that we are seeing where we are getting a file not
found error from the OS at connect time.

This should not be too big of a difference from the current path, it
just forces some of the work to serialize before other work that might
depend on it. In practice this appears to happen anyway most of the time
so this should not be that big of a difference.

Bug: 84938
Change-Id: Ieb8c7bc21f76663fdf75879f792e75715103f494
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602328
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>