[ffx] Adding serialization to emulator engines
Part of the lifecycle of an emulator engine in ffx emu is to serialize
the configuration to disk once the instance is started, so that it can
be read from disk on future invocations of show() and shutdown().

This CL implements the serialization for those engines. It's a common
trait that each engine will implement, so each engine need not
reimplement anything. We include a test which serializes and then
deserializes to verify the process is lossless.

Change-Id: I3a109669d78e5a86e1764746941e702e9e082bdf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608481
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>