[input] Adds the dead key support into the input pipeline
This change instantiates the dead key handler for the input pipeline
used in the workstation product.

For the time being, however, this change does not work as it should.
There is a KeyEvent protocol change that will likely be needed to make
things work smoothly.  See discussion thread at:
https://groups.google.com/a/fuchsia.dev/g/ui-input-dev/c/ITYKvbJS6_o

Bug: 75723
Change-Id: Ifa041ffd92098e17cfca03e15ae120b5697c5002
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604051
Reviewed-by: Alice Neels <neelsa@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>