[fint] Add enable_temporary_go_cache static field
We'll set this field on incremental builders instead of adding
`gocache_dir=getenv("TMPDIR") + "/gocache"` to each builder's GN args,
which is fragile because it assumes that $TMPDIR is set, which is not
the case on some developer workstations. Relying on Go to find the
temporary directory means we'll fall back to /tmp if there's no $TMPDIR
env var set.

Change-Id: I8a74365e13bebf124341515c36d74e5839678e3b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608741
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>