[ffx][emu] Add QemuBasedEngine trait and stage files
Adding a trait, QemuBasedEngine, which collects the common
methods used by all engines that use QEMU as the emulator platform.

In this initial CL, the method to copy the guest files into an
instance directory for the emulator engine is added. The guest_config
is returned with the updated paths.

Change-Id: Id106479041bdee457f00839f47aed1676755cac2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608477
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Daniel Dougherty <doughertyda@google.com>