[assembly] Remove duplicate GN groups for package sets
The GN groups in //build/images/BUILD.gn for each of the package sets
(base, cache, and universe) were still present, and duplicating logic in
//build/images/fuchsia/BUILD.gn and the assemble_system() template.

Further, the derived packages in meta_packages were being created twice,
once in //build/images/BUILD.gn and again by assemble_system() in
//build/images/fuchsia/BUILD.gn.

This removes the now-redundant packages and groups in
//build/images/BUILD.gn in lieu of newly created groups by
assemble_system(), which serve the same purpose.

While doing this, it was discovered that the fuchsia base package set
didn't include the policy_labels.swd check, and so those deps weren't
making it into the base package set.

This is change 3 of 5 for the soft transition for bug 81290

Bug: 81290
Bug: 82779
Bug: 82781

Change-Id: I9d880afc50711c09f4144838a02e02e38373e076
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607168
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>