[fshost] move data, blob, and pkgfs to core config
This moves the data, blob, and pkgfs capabilities, and all capabilities
that route to those filesystems under the hood, to the core_fshost.cml
config. This configuration is not included on bootfs, causing those
directories to not get routed when they will not be available.

This also adds these routes to the verify_routes allowlist, adding a new
allowlist just for bootfs so we continue to verify these routes on
non-bootfs-only builds. The verify_routes tool is extended to take
multiple allowlists so we can variably add this one.

Bug: 39588
Test: CQ
Change-Id: I05806ac345aa1c6d21ccd1fb69753c5f3b045fcd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569004
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>