[gidl] Add gidl-format host tool
This CL adds a formatter for .gidl files called gidl-format, integrates
it with fx format-code, and formats all existing .gidl files with it.

The formatter is a completely separate program from GIDL, but it uses
the same lexer: text/scanner from the Go standard library. Like gofmt,
it does not attempt to split or repack lines except in a few special
cases. The style is based based on the way we were already writing .gidl
files, and leans towards the gofmt style for Go literals.

Special comments `// gidl-format off` and `// gidl-format on` can be
used to disable the formatter in parts of a file.

The name gidl-format (not gidlfmt) was chosen to match fidl-format. Its
command-line interface is also modeled after fidl-format, and it
respects the rubric (https://fuchsia.dev/fuchsia-src/concepts/api/cli):

    Usage: gidl-format [-i] [<file>...]

    Format GIDL syntax in each <file>, or from standard input if no files are given.

    Options:
      -i    Formats file in place

Test: fx test gidl_format_tests
Change-Id: I1d420127f1ea81278cf0b6da583c905b84393654
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607972
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>