[plasa] Hook up the plasa tests into the SDK tests
I forgot to add this stanza earlier, leading to the plasa unit tests
not running on the builders. This allowed a broken test to slip in.
This change fixes the issue.

Also needed to fix up paths to the test files for the build to become
green on core.x64-release and other core.* builders.

Bug: 88914
Change-Id: I24d5a37de2bdb32f6b318ddb13314f8d3eca5e56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608472
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>