[wlan][sim] Support global DDK device functions
Previously the fake device manager of the sim environment required that
all devices be added by calling the fake device manager. This presented
some limitations for designs that did not necessarily want to integrate
with this device management framework. Specifically in this case the
network device component of the components library which aims to be a
generic library implementation without ties to any specific driver or
device management framework.

This change moves away from this requirement by implementing
device_add_from_driver and device_async_remove so that calls to
the global DDK functions device_add and device_async_remove will be
routed to FakeDevMgr. This still supports multiple FakeDevMgr by
recognizing which FakeDevMgr either the parent or the device is
associated with. This required some changes in how devices are
represented but the external interface presented by FakeDevMgr remains
mostly the same. Iteration of devices and DeviceId representations are
slightly different but should not cause major difficulties.

Additionally devices are stored in a linear structure so that the last
added deviced can be retrieved. This helps ensure that the correct
device is retrieved after adding it. There were instances when the
correct device was only retrieved because of a fortunate hash ordering.

Test: sim_device_test brcmfmac_sim_test iwlwifi_test
Change-Id: I43c9f5c4daffd856a859acad4f0fb96849ed4b65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597163
Reviewed-by: John Sheu <sheu@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>