[blobfs] Fix incorrect Merkle tree offset.
Due to BlockSizeType and the associated block size constants defined as
as uint32_t, multiplication of those values with other large uint32_t
variables led to an erroneous calculation of the Merkle tree offset when
the offset of the tree was > 4 GiB.

Expand BlockSizeType and associated constants from 32 to 64 bits.

Update tests where required as now certain comparisons require
signedness to be handled explicitly.

Bug: 88680
Test: fx test blobfs-unittest
Change-Id: Icada9863e01713141995338fcf008ddc84879732
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607761
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>