[password auth]Move minfs mock into disk_management.
The new account module depends on Minfs so its unit tests are also going to
need to use the MockMinfs that previously lived inside
account_manager::tests. The most logical home to me seems to be along
the trait but happy to talk about other options like a mocks module.

In the future we might want to move more of the account manager mocks
too but the fact that tests depend heavily on their private fields makes
that inconvenient at the moment.

Change-Id: I725d25e6a3be24ef20c50857371eed50ae052b3f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609083
Commit-Queue: Adam Lesinski <adamlesinski@google.com>
Reviewed-by: Adam Lesinski <adamlesinski@google.com>