[zircon][vm][ppb] DeleteLender in OnClose()
The DeleteLender() can happen on the entire range of
the contiguous VMO (the future ones where each on has a
PhysicalPageProvider).  We loan all the pages first to
make sure all pages are loaned, then the DeleteLender()
works on the whole range (including its assert that every
page is loaned).

Change-Id: I8716a3c5b786c4d1206084f8a9010ecc6653f4fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/603369
Commit-Queue: Dustin Green <dustingreen@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>