[audio_core][test] Fix HermeticFidelityTest for 44.1
Fix the incorrect calculation of source buffers, for SRC ratios
that require a fractional buffer length in order to translate
perfectly to the dest analysis buffer length of 65536 frames.
This is most clearly seen with test cases that include the
44.1k => 48k rate conversion, where SiNAD results would vary
by up to 12dB across runs.

Also, enable HermeticFidelityTest cases to specify that only a
single frequency should be tested. This will be most useful for
test cases that are generally frequency-independent, such as
noise-floor for uint8, int16 and int24-in-32 input formats.

Both updates are used in upcoming CL that adds the test suite
audio-core-fidelity-test.

Change-Id: I3e4632d45c4b17f6f2d54192de024a5b7fadd321
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590488
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>