[cm][storage] Wait for done signal in integ tests
The done signal indicates when our mock component has
completed writing to the provided directory. While waiting for
the signal is not necessary for the assertions made in the
test cases, the sender end of the signal in the mock panics if
the recipient end has been closed.

Change-Id: I868c8472e0a63401ec84732778391d7a8d2043eb
Fixed: 89079
Multiply: storage_admin_integration_test
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609006
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>