[debugger] Check stack for stopped thread.
Previously there were two functions to check for a stopped thread, one
that required a frame and one that didn't. However, all callers required
the thread be stopped with a frame so users of the former were
incorrect.

This converts all callers to use the "WithFrame" variant, and enhances
the checking for "stopped in exception."

Bug: 89073
Change-Id: I86f3100588ab16cf1d28a8bb4a1abd1da43e3233
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609465
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>