[assembly] Add destination field to assembly inputs
Some of the assembly inputs come from the root fuchsia directory, and
are therefore prefixed with ../../. In order to consolidate all the
artifacts under a single directory, we are adding a destination field
that strips the prefix.

The Product Config will then need to be updated on the consumer-side of
this file to also strip the prefix.

Bug: 87926
Change-Id: I523abc5dbc9a5ce4138aa39f545d1f7765f9daab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609467
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>