[network][explicit] Initial commit
Add `ResultExt` trait with `ok_checked` method. `ok_checked` is like
`Result::ok`, except that it takes a type parameter which must be equal
to the `Result`'s error type in order for a method call to compile. This
allows code to convert a `Result` to an `Option` in an ergonomic way
while complying with the Netstack team's policy on discarding type
information.

Change-Id: I6da381081b567f6b9b6d17db99669305968112f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608429
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>