[acpica] Only support S5 state from AcpiOsEnterSleep.
AcpiOsEnterSleep is used by ACPICA code to validate whether we want to
transition to a particular sleep state. Since the ACPICA functions used
to transition S-states require interrupts to be disabled and in Fuchsia
ACPICA runs in usermode and we don't expose a mechanism for it to
disable interrupts, we do not want to allow S-state transitions except
shutdown (sleep state S5) where cache corruption does not matter.

When we do support other S-state transitions we will do so by jumping to
the kernel to disable interrupts and make the transition so this
function will not be called.

Change-Id: Ie047f9d229f28a09281e3359c3077f9d6f6a02f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605221
Commit-Queue: Sarah Pham <smpham@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>