[ffx] Add deserialization to ffx emu engines
My previous commit added serialization to emulator engines, such that
the engine configuration is written to disk when startup succeeds. While
we confirmed the serialization is correct, it needs to be read back from
disk to be useful.

This CL adds the deserialization step, which is used by shutdown and
will be used by show (and maybe list). Some adjustments to serialization
were needed to enable deserializing before we know the type of engine,
and some error handling was added to both sides to make the interface
cleaner.

Change-Id: I1662d69ca846dad61711d48191e2905c29c5d7f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609322
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>