[audio][test] Adjust fidelity frequencies
In our fidelity testing, we use sinusoid input signals to characterize
a pipeline's frequency response and signal/noise. We measure frequency
response and signal/noise at specific frequencies, enforcing that the
results meet our expectations. Our set of test frequencies span from
subsonic (including 0hz) through the audible spectrum into the
ultrasonic range.

We get more accurate measurements because we avoid "windowing" the
output. This is possible because we use specific test frequencies that
lead to an integral number of output wavelengths in our power-of-two
buffer. We have discovered a side-effect of this test optimization.

Test frequencies that share many common factors with 48000 are prone
to slightly inaccurate results (+/- a few dB) compared to the rest of
the spectrum. Changing the frequency by even 1hz in either direction
leads to results that are much more consistent.

This CL changes our test frequencies to avoid frequencies that are
multiples of 375 (the non-power-of-two factor of 48000), specifically
21khz and 31.5 khz.

We still believe this no-windowing optimization is the best approach,
this slight workaround notwithstanding.

Change-Id: Iebec83ff5c74d3d63d81e9bb49af23db7ad3d20b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609979
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>