[network] Introduce network-basic
Introduce a basic version of the network realm that is, for now, the
same as the default version. In the future, when the default network
realm exposes virtualization functionality from netcfg, the
network-basic realm will instead have a stripped down version of netcfg.

Also, introduce a core shard specific to netcfg-advanced that will
expose this virtualization functionality, which for now is empty to
enable a soft transition.

This is step 1 of 3 for the soft transition tracked in fxbug.dev/87819.

Bug: 87819
Change-Id: I6e336fc770a4cb64e6f3341f86fef83063fc5a7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610042
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>