[fdio] Ignore MSG_NOSIGNAL
Commit 049ac5352f60b670f3bcac046731fc85eedbd290 defined this flag but
neglected to account for the fact that its definition would cause
downstream consumers to pass it to functions on the send path. Since
signals aren't implemented in Fuchsia, we can safely discard this flag
if it is passed.

Update various tests to pass this flag when it is defined, rather than
when __Fuchsia__ is not defined.

Fixed: b/207057734
Change-Id: I3671f4d27b4a4da385e3c8f928381cf2027947b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610383
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Neethu Joseph <neethujoseph@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>