[ffx][driver][lsblk] Add lsblk subcommand
The fuchsia-vfs-watcher library was also updated to work host-side. A
macro which in fuchsia-zircon-status was made public in order to make
this work, and a struct that was defined in fdio_sys was copied to avoid
dependencies that only work on fuchsia.

This doesn't implement support for lsblk subcommands - those should be
implemented as separate commands.

Tested: ffx driver lsblk
Change-Id: I67843808741484bfb79779bcc435304ecff52df6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610022
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>