[kernel][vm] VmMapping protection ranges
Changes the VmMapping object to support multiple protection regions,
instead of requiring a single protection type for the whole mapping.

Previously having many different protection ranges would require one
VmMapping per range, with operations on either the VMAR or the backing
VMO needing to iterate all these mappings. Not splitting into multiple
VmMapping can greatly improve the performance of these operations in
scenarios where there are multiple protection types.

As the common case is that there is one protection type for a mapping
this change aims to maintain performance for this case. This is done
so by short circuiting the protection region list if its empty and
still supporting one protection domain in a VmMapping without needing
additional allocations.

With this the penalty of this change is that all mappings now have an
extra two words, that represent the wavl tree, and so memory usage is
increased slightly.

The benefits are
 * Once protections happen a much smaller data structure, the
   ProtectNode, is allocated instead of a full VmMapping, thus reducing
   memory usage.
 * <=1% measurable performance impact to existing mapping
   microbenchmarks.
 * The previously mentioned benefits of mappings with lots of
   protections gaining a large performance improvement.

Bug: 86695

Change-Id: Ib7b32887506b65bf562ad893a5edeba951b2e073
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597986
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Kevin Marshall <marshallk@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>