[ffx][repo] Use SSH_CONNECTION for non-tunneled connections
Ideally we'd prefer connections from the target device to the ffx
repository to go through the ffx tunnel, since that would transparently
work when tunneling through a local device to a remote workstation.
Unfortunately the native ffx tunnel has a significant bug
http://fxbug.dev/87439, which can cause resolutions to fail.

To avoid that bug, this patch changes the repository configuration to
only create a tunnel if the server is running on a loopback address.
Otherwise, we'll instead tell the device to talk to the host address as
seen by the environment variable $SSH_CLIENT. This matches how
scripts like `fx serve-updates` configure devices to talk to the pm
server.

The downside of this approach is that it does not transparently work
well with scripts like `fx serve-remote`. Those scripts will need to be
updated in order to forward the repository address at the ssh level in
order for it to work.

Bug: 87439
Fixed: 88761

Change-Id: Ib778be33f3eafebba5576a16c30d3afc982a2dd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605986
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: James Tucker <raggi@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>