[ffx] Implement get_all_instances and ffx emu list
This CL adds the implementation for the get_all_instances function in
the ffx emu engines module. This function is currently used in the
"list" subcommand, to provide a listing of all of the current instances;
and also in the "shutdown" subcommand when the user passes in the
"--all" flag, to clean up all instances on the system.

Change-Id: Ic91b52c5f4f2b20fdc05cfee8afa7c04f68a2bd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610369
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>