[driver_manager] Append driver-related items to mexec data ZBI
There are ZBI items that the driver framework expects to be present in
the data ZBI that are not inspected by the kernel. Accordingly, now as
the canonical mexec caller, it would be more reasonable for
driver_manager to be responsible for appending these items before
mexec'ing (instead of this happening within the context of
`zx_system_mexec_get_payload()`).

Compatibility-wise, it should be safe to make the userspace and kernel
changes at the same time: an older version of userspace is not
guaranteed to work with a newer kernel - and, at worst, a newer
userspace that performs an mexec with an older kernel will have
duplicates of the driver metadata present in the ZBI.

Fixed: 89185
Change-Id: I0f603cfb768406f604d271dded7a0ba9285c0656
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610064
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>