[ffx] Implement a basic Show subcommand for emu
The last piece of the ffx emu interface is Show. It displays the details
for a given emulator instance, such as its status and the configuration
used to create it.

This CL implements a rudimentary Show subcommand for the Femu and Qemu
engine types. This subcommand just dumps the configuration to the
terminal using a debug string. It's not formatted in any particular way,
just uses Rust's default Pretty Print, but it serves as a starting point
for understanding what we still need to provide to make it a useful
tool.

Change-Id: I347126a8a7b7ad1ed5edaeebf2b7b5f25363b259
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610542
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>