[core][test] Avoid mappings being merged
There is a chance that when randomly placing mappings in a VMAR the
kernel happens to place a mapping in such a way that it is a candidate
for merging. This has been possible for a long time, but required the
two mappings have the same permission. Now they can have different
permissions and be merged, greatly increasing the odds of this flaking.

The fix in this CL is to map in every second page of the VMO, ensuring
that mappings can never be merged as they are not contiguous in VMO
object space.

Fixed: 89314
Change-Id: Ic6ea9aaff016ab42e4635d74c0f0b11769602fae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610568
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>